Add a timeout for the climb phase during RTH
If the climb phase lasts more than 10s and the aircraft is not
moving vertically at more than 10cm/s, consider the climb phase
finished and start going home while trying to continue climbing.

This will prevent an aircraft from getting stuck in the climb
phase forever, which is safer specially when the RTH is being
performed by the failsafe system.

References #1581